backend/layers/sql/requirements.txt
dnspython==1.16.0
mysql-connector-python==8.0.19
protobuf==3.6.1
six==1.14.0
